requirements.txt
importlib-metadata==1.6.0
jsonschema==3.2.0
python-telegram-bot==12.6.1
bs4==4.8.0
feedparser==5.2.1
requests==2.22.0
lxml=4.4.2